+++ example/pages/index.vue
@@ -11,19 +11,9 @@
     <section class="section">
       <div class="row">
         <div class="col-xs-12 col-sm-10 col-sm-offset-1">
-          <markdown-renderer type="url" :src="''"/>
+          <markdown-renderer src=""/>
         </div>
       </div>
     </section>
   </div>
 </template>
-
-<script>
-import markdownRenderer from '~/components/markdown/renderer';
-
-export default {
-  components: {
-    markdownRenderer
-  }
-};
-</script>