+++ example/pages/plugins/choicesjs-stencil.vue
@@ -11,7 +11,7 @@
     <section class="section">
       <div class="row">
         <div class="col-xs-12 col-sm-10 col-sm-offset-1">
-          <markdown-renderer type="url" :src="'choicesjs-stencil'"/>
+          <markdown-renderer src="choicesjs-stencil"/>
         </div>
       </div>
     </section>
@@ -47,7 +47,7 @@
             <h5 class="col-xs-12">
               {{ $t('views.plugins.plugin-choicesjs-stencil.example.type.multiple') }}
             </h5>
-            <choicesjs-stencil class="col-xs-12 choicesjs-stencil multiline"
+            <choicesjs-stencil class="col-xs-12 choicesjs-stencil choicesjs-stencil--multiline"
                 type="multiple"
                 name="multiple"
                 v-pre/>
@@ -64,19 +64,11 @@
           <h6 v-if="type">
             {{ $t(`views.plugins.plugin-choicesjs-stencil.example.result.${ type }`) }}
           </h6>
-          <div v-if="type === 'form'">
+          <div>
             <div class="results__item text-ellipsis" v-for="value in values" :key="value.id">
               <span class="results__item--type">({{ value.name }})</span>
               <span class="results__item--value" :title="transformFormValue(value.value)">
                 {{ transformFormValue(value.value) }}
-              </span>
-            </div>
-          </div>
-          <div v-else>
-            <div class="results__item text-ellipsis" v-for="event in events" :key="event.id">
-              <span class="results__item--type">({{ event.name }})</span>
-              <span class="results__item--value" :title="transformEventData(event.data)">
-                {{ transformEventData(event.data) }}
               </span>
             </div>
           </div>
@@ -87,16 +79,13 @@
 </template>
 
 <script>
-import markdownRenderer from '~/components/markdown/renderer';
-import { choices, createChoiceTemplate, createItemTemplate, events, getFormValues }
-  from '~/services/choicesjs-stencil';
+import { choices, createChoiceTemplate, createItemTemplate, getFormValues } from '~/services/choicesjs-stencil';
 
 export default {
   data() {
     return {
       type: '',
-      values: [],
-      events: []
+      values: []
     };
   },
   methods: {
@@ -117,11 +106,13 @@ export default {
     form.addEventListener('submit', (event) => {
       this.type = 'form';
       this.values = getFormValues(form);
-      this.events = [];
 
       event.preventDefault();
       event.stopImmediatePropagation();
     });
+
+    selectText.removeItems = true;
+    selectText.removeItemButton = true;
 
     selectMultiple.choices = choices;
     selectMultiple.maxItemCount = MAX_ITEMS;
@@ -143,22 +134,6 @@ export default {
       select.addItemText = (value) => this.$t('views.plugins.plugin-choicesjs-stencil.example.config.add-item', { value });
       select.maxItemText = (maxItemCount) => this.$t('views.plugins.plugin-choicesjs-stencil.example.config.max-items', { maxItemCount });
     });
-
-    [ selectText, selectSingle, selectMultiple ].forEach((select) => {
-      events.forEach((event) => {
-        select.addEventListener(event, (data) => {
-          this.type = 'event';
-          this.values = [];
-          this.events.push({
-            name: event,
-            data: data.detail
-          });
-        });
-      });
-    });
-  },
-  components: {
-    markdownRenderer
   }
 };
 </script>