+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.1
+
+- Fixed style of `choicesjs-stencil` package.
+- Updated example.
+  - Replaced markdown startup loading with `raw-loader`.
+  - Replaced markdown renderer with `vue-showdown` library.
+  - Removed the events section of ChoicesJS Stencil page.
+- Updated linter configurations.
+
 # 1.0.0
 
 - Initial version of `@adidas/yarn-design-system-plugins`.