+++ packages/choicesjs-stencil/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+
+- Fixed hidden input style.
+- Replaced `.multiline` class with `.choicesjs-stencil--multiline`.
+
 # 1.1.0
 
 - Updated peer dependency `choicesjs-stencil` to `^1.2`.